+++ src/main/java/com/kbds/itamserveradmin/domain/asset/entity/AssetNotice.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.apache.tomcat.util.http.parser.ContentRange;
 
 import javax.persistence.*;
