+++ src/main/java/com/kbds/itamserveradmin/domain/contract/entity/ContractRecord.java
@@ -1,4 +1,4 @@
-package com.kbds.itamserveradmin.domain.contract.repository.entity;
+package com.kbds.itamserveradmin.domain.contract.entity;
 
 import com.kbds.itamserveradmin.domain.user.entity.User;
 import com.kbds.itamserveradmin.global.entity.BaseEntity;
