+++ src/main/java/com/kbds/itamserveradmin/domain/contract/entity/OpStatus.java
@@ -1,0 +1,5 @@
+package com.kbds.itamserveradmin.domain.contract.entity;
+
+public enum OpStatus {
+    IN_OPERATION, IN_DISPOSAL
+}
