--- src/main/java/com/kbds/itamserveradmin/domain/contract/repository/entity/OpStatus.java
@@ -1,5 +1,0 @@
-package com.kbds.itamserveradmin.domain.contract.repository.entity;
-
-public enum OpStatus {
-    IN_OPERATION, IN_DISPOSAL
-}
