+++ src/main/java/com/kbds/itamserveradmin/domain/contract/service/ContractService.java
@@ -1,4 +1,4 @@
-package com.kbds.itamserveradmin.domain.contract.repository.service;
+package com.kbds.itamserveradmin.domain.contract.service;
 
 import org.springframework.stereotype.Service;
