--- src/main/java/com/kbds/itamserveradmin/domain/corperation/controller/CooperationController.java
@@ -1,4 +1,0 @@
-package com.kbds.itamserveradmin.domain.cooperation.controller;
-
-public class CooperationController {
-}
