--- src/main/java/com/kbds/itamserveradmin/domain/corperation/repository/CorporationRepository.java
@@ -1,8 +1,0 @@
-package com.kbds.itamserveradmin.domain.cooperation.repository;
-
-import com.kbds.itamserveradmin.domain.cooperation.entity.Cooperation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CooperationRepository extends JpaRepository<Cooperation,String> {
-
-}
