--- src/main/java/com/kbds/itamserveradmin/domain/corperation/service/CooperationService.java
@@ -1,4 +1,0 @@
-package com.kbds.itamserveradmin.domain.cooperation.service;
-
-public interface CooperationService {
-}
