--- src/main/java/com/kbds/itamserveradmin/domain/corperation/service/CooperationServiceImpl.java
@@ -1,4 +1,0 @@
-package com.kbds.itamserveradmin.domain.cooperation.service;
-
-public class CooperationServiceImpl implements CooperationService{
-}
