+++ src/main/java/com/kbds/itamserveradmin/domain/corporation/controller/CorporationController.java
@@ -1,4 +1,4 @@
 package com.kbds.itamserveradmin.domain.corporation.controller;
 
-public class CooperationController {
+public class CorporationController {
 }
