+++ src/main/java/com/kbds/itamserveradmin/domain/corporation/repository/CorporationRepository.java
@@ -3,6 +3,6 @@ package com.kbds.itamserveradmin.domain.corporation.repository;
 import com.kbds.itamserveradmin.domain.corporation.entity.Corporation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CooperationRepository extends JpaRepository<Corporation,String> {
+public interface CorporationRepository extends JpaRepository<Corporation,String> {
 
 }
