--- src/main/java/com/kbds/itamserveradmin/domain/corporation/service/CooperationServiceImpl.java
@@ -1,4 +1,0 @@
-package com.kbds.itamserveradmin.domain.corporation.service;
-
-public class CooperationServiceImpl implements CooperationService{
-}
