+++ src/main/java/com/kbds/itamserveradmin/domain/corporation/service/CorporationService.java
@@ -1,4 +1,4 @@
 package com.kbds.itamserveradmin.domain.corporation.service;
 
-public interface CooperationService {
+public interface CorporationService {
 }
