+++ src/main/java/com/kbds/itamserveradmin/domain/corporation/service/CorporationServiceImpl.java
@@ -1,0 +1,4 @@
+package com.kbds.itamserveradmin.domain.corporation.service;
+
+public class CorporationServiceImpl implements CorporationService {
+}
