+++ src/main/java/com/kbds/itamserveradmin/domain/user/entity/User.java
@@ -3,7 +3,7 @@ package com.kbds.itamserveradmin.domain.user.entity;
 
 import com.kbds.itamserveradmin.domain.asset.entity.AssetNotice;
 import com.kbds.itamserveradmin.domain.assetRequest.entity.AssetRequest;
-import com.kbds.itamserveradmin.domain.contract.repository.entity.ContractRecord;
+import com.kbds.itamserveradmin.domain.contract.entity.ContractRecord;
 import com.kbds.itamserveradmin.domain.corporation.entity.Corporation;
 import com.kbds.itamserveradmin.domain.department.entity.Department;
 import com.kbds.itamserveradmin.global.entity.BaseEntity;
